Fetch only when we know we need to fetch
Git rev-list does NOT do a fetch, nor does it get the remote branch
points, as it can't. The reason it can't is that do list or count the
commits between the two it needs the whole history, which would be a
fetch. Git ls-remote ONLY gets the commit at the remote branch, not any
history. This means it does less work, so faster. We only need to update
if there is a difference, so we do that check first to save time. Only
then will we git fetch to get all the details that we want.